+++ client/src/pages/Home.js
@@ -2,7 +2,7 @@ import React, { useEffect } from "react";
 import { MDBCol, MDBContainer, MDBRow, MDBTypography } from "mdb-react-ui-kit";
 import { useDispatch, useSelector } from "react-redux";
 import { getTours } from "../redux/actions/tourActions";
-import {  setCurrentPage} from "../redux/features/tourSlice"
+import { setCurrentPage } from "../redux/features/tourSlice";
 import CardTour from "../components/CardTour";
 import Spinner from "../components/Spinner";
 import Pagination from "../components/Pagination";
@@ -33,8 +33,7 @@ const Home = () => {
     return <Spinner />;
   }
   return (
-    <MDBContainer className="pt-5"
-    >
+    <MDBContainer className="pt-5">
       <MDBRow className="mt-5">
         {tours.length === 0 && location.pathname === "/" && (
           <MDBTypography className="text-center mb-0" tag="h2">
@@ -48,26 +47,26 @@ const Home = () => {
           </MDBTypography>
         )}
 
-
         <MDBRow className="g-2">
-
           <MDBCol size="12" lg="10" xl="8">
-            {tours &&
-              tours.map((item) => <CardTour key={item._id} {...item} />)}
+            <MDBRow>
+              {tours &&
+                tours.map((item) => <CardTour key={item._id} {...item} />)}
+            </MDBRow>
           </MDBCol>
-          <MDBCol size="12" lg="2" xl="4" >
-            <h3 className="me-auto">
-              News Feed
-            </h3>
+          <MDBCol size="12" lg="2" xl="4">
+            {tours.length > 0 && location.pathname === "/" && (
+              <h3 className="me-auto">News Feed</h3>
+            )}
             <div className="mt-2">
-              <NewFeed/>
+              {tours &&
+                tours
+                  .map((item) => <NewFeed key={item._id} {...item} />)
+                  .reverse()
+                  .splice(0, 4)}
             </div>
-              
           </MDBCol>
-
         </MDBRow>
-
-
       </MDBRow>
       {tours.length > 0 && !searchQuery && (
         <Pagination
