+++ client/src/pages/Register.js
@@ -12,7 +12,7 @@ import {
 import { Link, useNavigate } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 import { toast } from "react-toastify";
-import { register } from "../redux/features/authSlice";
+import { register } from "../redux/actions/authActions";
 
 const initialState = {
   firstName: "",
