+++ client/src/redux/actions/authActions.js
@@ -1,0 +1,45 @@
+import { createAsyncThunk } from "@reduxjs/toolkit";
+import * as api from "../api";
+
+export const login = createAsyncThunk(
+  "auth/login",
+  async ({ formValue, navigate, toast }, { rejectWithValue }) => {
+    try {
+      const response = await api.signIn(formValue);
+      toast.success("Login Successfully");
+      navigate("/");
+      console.log(response.data);
+      return response.data;
+    } catch (err) {
+      return rejectWithValue(err.response.data);
+    }
+  }
+);
+
+export const register = createAsyncThunk(
+  "auth/register",
+  async ({ formValue, navigate, toast }, { rejectWithValue }) => {
+    try {
+      const response = await api.signUp(formValue);
+      toast.success("Register Successfully");
+      navigate("/");
+      return response.data;
+    } catch (err) {
+      return rejectWithValue(err.response.data);
+    }
+  }
+);
+
+export const googleSignIn = createAsyncThunk(
+  "auth/googleSignIn",
+  async ({ result, navigate, toast }, { rejectWithValue }) => {
+    try {
+      const response = await api.googleSignIn(result);
+      toast.success("Google Sign-in Successfully");
+      navigate("/");
+      return response.data;
+    } catch (err) {
+      return rejectWithValue(err.response.data);
+    }
+  }
+);
